refactor: reorganize history management by moving functions to dedicated modules and updating imports
